requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Pillow==8.2.0
pony==0.7.14
psycopg2==2.8.6
requests==2.25.1
six==1.15.0
urllib3==1.26.4
vk-api==11.9.1
